+++ History.md
@@ -3,7 +3,9 @@
   * Fix Cobertura report for file-system instrumentation (https://github.com/tntim96/JSCover/pull/50)
   * Fix continue label handling (https://github.com/tntim96/JSCover/issues/51)
   * Fix MIME image mappings
-  * Add tests for headers in proxy for '.js' files (https://github.com/tntim96/JSCover/issues/49)
+  * Fix proxy report for viewing source (https://github.com/tntim96/JSCover/issues/53)
+  * Internal: Upgrade to HTMLUnit 2.12
+              Add tests for headers in proxy for '.js' files (https://github.com/tntim96/JSCover/issues/49)
 
 0.2.5 / 2013-03-04
 ==================