+++ History.md
@@ -1,4 +1,4 @@
-0.2.6 / 2013-03-??
+0.2.6 / 2013-03-11
 ==================
   * Fix Cobertura report for file-system instrumentation (https://github.com/tntim96/JSCover/pull/50)
   * Fix continue label handling (https://github.com/tntim96/JSCover/issues/51)